refactor: move theme and globalstyles to layout component
