+++ src/pages/index.tsx
@@ -1,22 +1,16 @@
 import React from "react"
 import { Link } from "gatsby"
-import { ThemeProvider } from "styled-components"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import { theme } from "../theme/index"
 import Button from "../components/button"
-import GlobalStyle from "../styles/Global.js"
 
 const IndexPage = () => (
-  <ThemeProvider theme={theme}>
-    <GlobalStyle />
-    <Layout>
-      <SEO title="Home" />
-      <Button>Test test</Button>
-      <Link to="/page-2/">Go to page 2</Link>
-    </Layout>
-  </ThemeProvider>
+  <Layout>
+    <SEO title="Home" />
+    <Button>Test test</Button>
+    <Link to="/page-2/">Go to page 2</Link>
+  </Layout>
 )
 
 export default IndexPage
